more protocol checks. try to fix 'many pos packets' kick

+++ protocol/src/main/java/net/md_5/bungee/protocol/packet/extra/ConfirmTransaction.java
@@ -17,7 +17,7 @@ public class ConfirmTransaction extends DefinedPacket
 {
 
     byte window;
-    short action;
+    int action;
     boolean accepted;
 
     @Override
